The entire `HandshakeState` object now is a state machine. It will validate correctness by determining whether it is okay to read or write, or whether the handshake is completed or not, and will not allow you to read, write, or finalize (split) until all conditions are met. Plus a few extra changes too.
The Custom handshake pattern was removed. I see little value in keeping it, particularly since it is questionable how secure a "custom" pattern would be to begin with.

The read_message function now validates that the message length is less than the upper bound as provided for by the Noise specification.